requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
berserk==0.14.0
certifi==2026.2.25
charset-normalizer==3.4.5
Deprecated==1.3.1
discord.py==2.7.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
ndjson==0.3.1
pillow==12.1.1
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
